Fixed the code more
FIxed the code more and dropped the idea of using plugins, it was a bit problematic, so since it was just a calc it was ok. but it would be fun. and now using visual studio professional instead. and now we got the calc to work simply. but not tottally